requirements.txt
django==6.0.3
djangorestframework==3..17.1
django-cors-headers==4.9.0
requests==2.33.1